[python] move make_walker, make_apply_on_dirichlet_intersections
